Removed/renamed files for design smell-1
