+++ telegrambots-abilities/src/main/java/org/telegram/abilitybots/api/sender/DefaultSender.java
@@ -3,9 +3,6 @@ package org.telegram.abilitybots.api.sender;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.groupadministration.SetChatPhoto;
 import org.telegram.telegrambots.meta.api.methods.send.*;
-import org.telegram.telegrambots.meta.api.methods.stickers.AddStickerToSet;
-import org.telegram.telegrambots.meta.api.methods.stickers.CreateNewStickerSet;
-import org.telegram.telegrambots.meta.api.methods.stickers.UploadStickerFile;
 import org.telegram.telegrambots.meta.api.objects.File;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.User;
@@ -27,10 +24,6 @@ public class DefaultSender implements MessageSender {
 
   private DefaultAbsSender bot;
 
-  /*
-  *  Refactored code - Insufficient Modularization design smell refactored by moving the methods related to send functionality
-  * and extracting class DefaultSender1.
-   */
   public DefaultSender(DefaultAbsSender bot) {
     this.bot = bot;
   }
@@ -43,21 +36,6 @@ public class DefaultSender implements MessageSender {
   @Override
   public <T extends Serializable, Method extends BotApiMethod<T>> T execute(Method method) throws TelegramApiException {
     return bot.execute(method);
-  }
-
-  @Override
-  public Boolean addStickerToSet(AddStickerToSet addStickerToSet) throws TelegramApiException {
-    return bot.execute(addStickerToSet);
-  }
-
-  @Override
-  public Boolean createNewStickerSet(CreateNewStickerSet createNewStickerSet) throws TelegramApiException {
-    return bot.execute(createNewStickerSet);
-  }
-
-  @Override
-  public File uploadStickerFile(UploadStickerFile uploadStickerFile) throws TelegramApiException {
-    return bot.execute(uploadStickerFile);
   }
 
   @Override
@@ -109,5 +87,35 @@ public class DefaultSender implements MessageSender {
   @Override
   public Message sendDocument(SendDocument sendDocument) throws TelegramApiException {
     return bot.execute(sendDocument);
+  }
+
+  @Override
+  public Message sendPhoto(SendPhoto sendPhoto) throws TelegramApiException {
+    return bot.execute(sendPhoto);
+  }
+
+  @Override
+  public Message sendVideo(SendVideo sendVideo) throws TelegramApiException {
+    return bot.execute(sendVideo);
+  }
+
+  @Override
+  public Message sendSticker(SendSticker sendSticker) throws TelegramApiException {
+    return bot.execute(sendSticker);
+  }
+
+  @Override
+  public Message sendAudio(SendAudio sendAudio) throws TelegramApiException {
+    return bot.execute(sendAudio);
+  }
+
+  @Override
+  public Message sendVoice(SendVoice sendVoice) throws TelegramApiException {
+    return bot.execute(sendVoice);
+  }
+
+  @Override
+  public Message sendVideoNote(SendVideoNote sendVideoNote) {
+    return null;
   }
 }
