+++ telegrambots-abilities/src/main/java/org/telegram/abilitybots/api/sender/MessageSender.java
@@ -3,9 +3,6 @@ package org.telegram.abilitybots.api.sender;
 import org.telegram.telegrambots.meta.api.methods.BotApiMethod;
 import org.telegram.telegrambots.meta.api.methods.groupadministration.SetChatPhoto;
 import org.telegram.telegrambots.meta.api.methods.send.*;
-import org.telegram.telegrambots.meta.api.methods.stickers.AddStickerToSet;
-import org.telegram.telegrambots.meta.api.methods.stickers.CreateNewStickerSet;
-import org.telegram.telegrambots.meta.api.methods.stickers.UploadStickerFile;
 import org.telegram.telegrambots.meta.api.objects.File;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.User;
@@ -28,12 +25,6 @@ public interface MessageSender {
 
   <T extends Serializable, Method extends BotApiMethod<T>> T execute(Method method) throws TelegramApiException;
 
-  Boolean addStickerToSet(AddStickerToSet addStickerToSet) throws TelegramApiException;
-
-  Boolean createNewStickerSet(CreateNewStickerSet createNewStickerSet) throws TelegramApiException;
-
-  File uploadStickerFile(UploadStickerFile uploadStickerFile) throws TelegramApiException;
-
   Boolean setChatPhoto(SetChatPhoto setChatPhoto) throws TelegramApiException;
 
   java.io.File downloadFile(String path) throws TelegramApiException;
@@ -54,4 +45,15 @@ public interface MessageSender {
 
   Message sendDocument(SendDocument sendDocument) throws TelegramApiException;
 
+  Message sendPhoto(SendPhoto sendPhoto) throws TelegramApiException;
+
+  Message sendVideo(SendVideo sendVideo) throws TelegramApiException;
+
+  Message sendAudio(SendAudio sendAudio) throws TelegramApiException;
+
+  Message sendVoice(SendVoice sendVoice) throws TelegramApiException;
+
+  Message sendVideoNote(SendVideoNote sendVideoNote) throws TelegramApiException;
+
+  Message sendSticker(SendSticker sendSticker) throws TelegramApiException;
 }
